requirements.txt
Flask>=3.0,<4
Werkzeug>=3.0
requests>=2.31
gunicorn>=21.2
python-dotenv>=1.0
